+++ app/src/test/kotlin/com/adammcneilly/pokedex/list/PokemonListCardPaparazziTest.kt
@@ -1,7 +1,9 @@
-package com.adammcneilly.pokedex
+package com.adammcneilly.pokedex.list
 
+import com.adammcneilly.pokedex.BasePaparazziTest
 import com.adammcneilly.pokedex.displaymodels.ImageDisplayModel
 import com.adammcneilly.pokedex.displaymodels.PokemonDisplayModel
+import com.adammcneilly.pokedex.list.PokemonListCard
 import com.adammcneilly.pokedex.models.PokemonType
 import org.junit.Test
