+++ app/src/test/snapshots/images/com.adammcneilly.pokedex.list_PokemonListCardPaparazziTest_renderDefault[useDarkTheme=false].png
